@@ -1,11 +1,14 @@
 {
     "name": "sakai-vue",
     "version": "4.3.0",
+    "type": "module",
     "scripts": {
+        "postinstall": "patch-package",
         "dev": "vite",
         "build": "vite build",
         "preview": "vite preview",
-        "lint": "eslint --fix . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
+        "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs",
+        "lint:fix": "eslint --fix . --ext .vue,.js,.jsx,.cjs,.mjs"
     },
     "dependencies": {
         "@primeuix/themes": "^1.2.3",
@@ -22,30 +25,29 @@
     },
     "devDependencies": {
         "@primevue/auto-import-resolver": "^4.3.1",
-        "@rushstack/eslint-patch": "^1.8.0",
         "@tsconfig/node22": "^22.0.2",
         "@types/eventemitter2": "^2.2.1",
         "@types/leaflet": "^1.9.20",
         "@types/node": "^24.3.0",
+        "@typescript-eslint/eslint-plugin": "8.46.0",
+        "@typescript-eslint/parser": "8.46.0",
         "@vitejs/plugin-vue": "^6.0.1",
         "@vue/eslint-config-prettier": "^10.2.0",
-        "@vue/eslint-config-typescript": "^14.6.0",
         "@vue/tsconfig": "^0.8.1",
         "autoprefixer": "^10.4.19",
-        "eslint": "^9.34.0",
+        "eslint": "^9.37.0",
+        "eslint-plugin-prettier": "5.5.4",
         "eslint-plugin-vue": "^10.4.0",
         "jiti": "^2.5.1",
+        "patch-package": "8.0.1",
         "postcss": "^8.4.40",
         "prettier": "^3.2.5",
         "sass": "^1.55.0",
         "tailwindcss": "^3.4.6",
-        "typescript": "^5.9.2",
+        "typescript": "5.9.3",
         "unplugin-vue-components": "^0.27.3",
         "vite": "^7.1.7",
         "vite-plugin-vue-devtools": "^8.0.1",
         "vue-tsc": "^3.0.6"
-    },
-    "overrides": {
-        "chalk": "<5.6.2"
     }
 }
