@@ -11,133 +11,134 @@ const router = createRouter({
         {
           path: '/',
           name: 'dashboard',
-          component: () => import('@/views/Dashboard.vue')
+          component: () => import('@/views/Dashboard.vue'),
         },
         {
-          path: '/map',
+          path: '/map/:id',
           name: 'map',
-          component: () => import('@/!custom/views/pages/Map.vue')
+          component: () => import('@/_custom/views/pages/Map.vue'),
+          props: true, // optional: passes route params as props to the component
         },
         {
           path: '/uikit/formlayout',
           name: 'formlayout',
-          component: () => import('@/views/uikit/FormLayout.vue')
+          component: () => import('@/views/uikit/FormLayout.vue'),
         },
         {
           path: '/uikit/input',
           name: 'input',
-          component: () => import('@/views/uikit/InputDoc.vue')
+          component: () => import('@/views/uikit/InputDoc.vue'),
         },
         {
           path: '/uikit/button',
           name: 'button',
-          component: () => import('@/views/uikit/ButtonDoc.vue')
+          component: () => import('@/views/uikit/ButtonDoc.vue'),
         },
         {
           path: '/uikit/table',
           name: 'table',
-          component: () => import('@/views/uikit/TableDoc.vue')
+          component: () => import('@/views/uikit/TableDoc.vue'),
         },
         {
           path: '/uikit/list',
           name: 'list',
-          component: () => import('@/views/uikit/ListDoc.vue')
+          component: () => import('@/views/uikit/ListDoc.vue'),
         },
         {
           path: '/uikit/tree',
           name: 'tree',
-          component: () => import('@/views/uikit/TreeDoc.vue')
+          component: () => import('@/views/uikit/TreeDoc.vue'),
         },
         {
           path: '/uikit/panel',
           name: 'panel',
-          component: () => import('@/views/uikit/PanelsDoc.vue')
+          component: () => import('@/views/uikit/PanelsDoc.vue'),
         },
 
         {
           path: '/uikit/overlay',
           name: 'overlay',
-          component: () => import('@/views/uikit/OverlayDoc.vue')
+          component: () => import('@/views/uikit/OverlayDoc.vue'),
         },
         {
           path: '/uikit/media',
           name: 'media',
-          component: () => import('@/views/uikit/MediaDoc.vue')
+          component: () => import('@/views/uikit/MediaDoc.vue'),
         },
         {
           path: '/uikit/message',
           name: 'message',
-          component: () => import('@/views/uikit/MessagesDoc.vue')
+          component: () => import('@/views/uikit/MessagesDoc.vue'),
         },
         {
           path: '/uikit/file',
           name: 'file',
-          component: () => import('@/views/uikit/FileDoc.vue')
+          component: () => import('@/views/uikit/FileDoc.vue'),
         },
         {
           path: '/uikit/menu',
           name: 'menu',
-          component: () => import('@/views/uikit/MenuDoc.vue')
+          component: () => import('@/views/uikit/MenuDoc.vue'),
         },
         {
           path: '/uikit/charts',
           name: 'charts',
-          component: () => import('@/views/uikit/ChartDoc.vue')
+          component: () => import('@/views/uikit/ChartDoc.vue'),
         },
         {
           path: '/uikit/misc',
           name: 'misc',
-          component: () => import('@/views/uikit/MiscDoc.vue')
+          component: () => import('@/views/uikit/MiscDoc.vue'),
         },
         {
           path: '/uikit/timeline',
           name: 'timeline',
-          component: () => import('@/views/uikit/TimelineDoc.vue')
+          component: () => import('@/views/uikit/TimelineDoc.vue'),
         },
         {
           path: '/pages/empty',
           name: 'empty',
-          component: () => import('@/views/pages/Empty.vue')
+          component: () => import('@/views/pages/Empty.vue'),
         },
         {
           path: '/pages/crud',
           name: 'crud',
-          component: () => import('@/views/pages/Crud.vue')
+          component: () => import('@/views/pages/Crud.vue'),
         },
         {
           path: '/documentation',
           name: 'documentation',
-          component: () => import('@/views/pages/Documentation.vue')
-        }
-      ]
+          component: () => import('@/views/pages/Documentation.vue'),
+        },
+      ],
     },
     {
       path: '/landing',
       name: 'landing',
-      component: () => import('@/views/pages/Landing.vue')
+      component: () => import('@/views/pages/Landing.vue'),
     },
     {
       path: '/pages/notfound',
       name: 'notfound',
-      component: () => import('@/views/pages/NotFound.vue')
+      component: () => import('@/views/pages/NotFound.vue'),
     },
 
     {
       path: '/auth/login',
       name: 'login',
-      component: () => import('@/views/pages/auth/Login.vue')
+      component: () => import('@/views/pages/auth/Login.vue'),
     },
     {
       path: '/auth/access',
       name: 'accessDenied',
-      component: () => import('@/views/pages/auth/Access.vue')
+      component: () => import('@/views/pages/auth/Access.vue'),
     },
     {
       path: '/auth/error',
       name: 'error',
-      component: () => import('@/views/pages/auth/Error.vue')
-    }
-  ]
+      component: () => import('@/views/pages/auth/Error.vue'),
+    },
+  ],
 });
 
 export default router;
